TI-Feature: power_management
TI-Tree: git://git.ti.com/~kristo/ti-linux-kernel/pm-linux-feature-tree.git
TI-Branch: pm-linux-3.8.y

* 'pm-linux-3.8.y' of git://git.ti.com/~kristo/ti-linux-kernel/pm-linux-feature-tree: (33 commits)
  staging: ti-soc-thermal: fix device removal
  staging: ti-soc-thermal: update OMAP5 extrapolation rules
  staging: ti-soc-thermal: introduce OMAP4430 extrapolation constants
  staging: ti-soc-thermal: Remove TC1/TC2 TODO (already done)
  staging: ti-soc-thermal: fix min/max TODO (already done)
  staging: ti-soc-thermal: update TODO list
  staging: ti-soc-thermal: Add get_trend support
  staging: ti-soc-thermal:Introduce ti_bandgap_get_trend function for OMAP5
  staging: ti-soc-thermal: Enable HISTORY_BUFFER Feature for OMAP5
  staging: ti-soc-thermal: Introduce HAS_HISTORY_BUFFER feature for bandgap
  staging: ti-soc-thermal: Modify update_interval r/w functions to incorporate the OMAP5 feature of COUNTER_DELAY.
  staging: ti-soc-thermal: Enable COUNTER_DELAY feature for OMAP5
  staging: ti-soc-thermal: Introduce HAS_COUNTER_DELAY feature for bandgap
  staging: ti-soc-thermal: Initialise counter_delay field for OMAP5 sensors
  staging: ti-soc-thermal: Add counter_delay_mask field to temp_sensor_registers struct
  arm: omap2plus_defconfig: enable TI bandgap driver
  staging: ti-soc-thermal: defer probe if cpufreq is not ready
  cpufreq: Add a get_current_driver helper
  arm: add bandgap DT entry for OMAP5
  ARM: OMAP3+: use cpu0-cpufreq driver in device tree supported boot(v4)
  ...

Conflicts:
	arch/arm/boot/dts/am33xx.dtsi
	arch/arm/boot/dts/omap5.dtsi
	arch/arm/configs/omap2plus_defconfig
	arch/arm/mach-omap2/omap-mpuss-lowpower.c
	arch/arm/mach-omap2/omap_hwmod_33xx_data.c

+++ arch/arm/mach-omap2/omap-mpuss-lowpower.c
@@ -98,7 +98,7 @@ static int default_finish_suspend(unsigned long cpu_state)
 static void dummy_cpu_resume(void)
 {}
 
-static void dummy_scu_prepare(unsigned int cpu_id, unsigned int cpu_state)
+static void dummy_scu_prepare(unsigned int cpu_id, u8 cpu_state)
 {}
 
 static struct cpu_pm_ops omap_pm_ops = {
